Added serializer-based class able to make deep copies of objects — CMemorySerializer. CGameState::init will work on the deep copy of StartInfo. Some casts to reduce warnings.

+++ lib/CRandomGenerator.h
@@ -26,7 +26,7 @@ public:
 	/// Seeds the generator with the current time by default.
 	CRandomGenerator() 
 	{
-		gen.seed(std::time(nullptr));
+		gen.seed((unsigned long)std::time(nullptr));
 	}
 
 	void seed(int value)
